rename from `stats` to `info`
