+++ docs/reference/cluster.asciidoc
@@ -104,7 +104,7 @@ include::cluster/nodes-reload-secure-settings.asciidoc[]
 
 include::cluster/nodes-stats.asciidoc[]
 
-include::cluster/http-stats.asciidoc[]
+include::cluster/info-http.asciidoc[]
 
 include::cluster/pending.asciidoc[]
