+++ rest-api-spec/src/main/resources/rest-api-spec/api/info.http.json
@@ -1,7 +1,7 @@
 {
-  "http.stats": {
+  "info.http": {
       "documentation": {
-          "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/cluster-http-stats.html",
+          "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/cluster-info-http.html",
           "description": "Returns statistical information about http status in the cluster."
       },
       "stability": "stable",
@@ -12,7 +12,7 @@
       "url": {
           "paths": [
               {
-                  "path": "/_http_stats",
+                  "path": "/_info/http",
                   "methods": ["GET"]
               }
           ]
